requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
starlette
fastapi
uvicorn
pydantic
sentence-transformers
langchain
langchain-community
chromadb
sentence-transformers
requests
streamlit
python-dotenv
openai
tqdm
